% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantmix-package.R
\docType{package}
\name{plantmix-package}
\alias{plantmix-package}
\alias{plantmix}
\title{The plantmix package}
\description{
plantmix is a package to study plant mixtures.
}
\section{Introduction}{

The main goal of the plantmix package is to provide a set of functions
for pain-free analysis of both varietal (intraspecific) and
crop (interspecific) mixtures, such as
optimizing experimental designs, simulating data sets, fitting
GMA-SMA and DBV-SBV models, computing diversity indices, etc.
Work is ongoing on DBV-SBV models for varietal mixtures, spatial heterogeneity, etc.
}

\section{Optimizing experimental designs}{

\itemize{
  \item \code{\link{getDesignBinaryVarMix}}, \code{\link{getDesignBinaryCropMix}}
  \item \code{\link{plotDesignVarMix}}, \code{\link{plotDesignCropMix}}
  \item \code{\link{plotDiallel}}, \code{\link{getMixtureList}}, \code{\link{getMixturesPerGeno}}
}
}

\section{Fitting GMA-SMA models}{

\itemize{
  \item \code{\link{mkZGMA}}, \code{\link{mkZSMA}}, \code{\link{mkAllZSMA}}
  \item \code{\link{fitGMASMA}}, \code{\link{summarizeGMASMAs}}
}
}

\section{Fitting DBV-SBV models}{

\itemize{
  \item \code{\link{pivotMixData2Long}}, \code{\link{pivotMixData2Wide}}
  \item \code{\link{mkZinterspe}}, \code{\link{fitDBVSBVinter}}
}
}

\section{Simulating mixture data}{

\itemize{
  \item \code{\link{simulDBVSBVinter}}
}
}

\section{Interaction indices}{

\itemize{
  \item \code{\link{RYM}}
  \item \code{\link{RY}}, \code{\link{RYT}}, \code{\link{LER}}, \code{\link{RYP}}
  \item \code{\link{CC}}, \code{\link{RII}}, \code{\link{RIInet}}
  \item \code{\link{estimRYMRep}}, \code{\link{estimRYRep}}
}
}

\section{Miscellaneous}{

\itemize{
  \item \code{\link{rmatnorm}}, \code{\link{vec}}, \code{\link{invvec}}, \code{\link{imageMat}}
  \item \code{\link{simulGenosDoseStruct}}, \code{\link{estimGRM}}
  \item \code{\link{paramBoot4TMB}}
  \item \code{\link{mixSowingWeight}}, \code{\link{nbSeedsToSownInPure}}
}
}

\author{
\strong{Maintainer}: Timothee Flutre \email{timothee.flutre@inrae.fr} (\href{https://orcid.org/0000-0003-4489-4782}{ORCID}) [contributor]

Other contributors:
\itemize{
  \item  INRAE [copyright holder, funder]
  \item Jerome Enjalbert \email{jerome.enjalbert@inrae.fr} (\href{https://orcid.org/0000-0003-0408-2253}{ORCID}) [contributor]
  \item Emma Forst \email{emma.forst@inra.fr} (\href{https://orcid.org/0000-0001-9557-3589}{ORCID}) [contributor]
  \item Maxence Remerand \email{maxence.remerand@ens-paris-saclay.fr} (\href{https://orcid.org/0009-0007-3217-5613}{ORCID}) [contributor]
  \item Jemay Salomon \email{jemaysalomon@gmail.com} (\href{https://orcid.org/0000-0003-4454-7916}{ORCID}) [contributor]
}

}
\keyword{internal}
