% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_RY.R
\name{estimRYRep}
\alias{estimRYRep}
\title{RY with "rep" effect}
\usage{
estimRYRep(
  data,
  colY,
  colIDstand,
  colIDfocal,
  colRep,
  mix2genos,
  conf_level = 0.95,
  plotDiag = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{colY}{column name specifying the response variable}

\item{colIDstand}{column name specifying the stand identifiers}

\item{colIDfocal}{column name for the focal identifiers}

\item{colRep}{column name specifying the replicate}

\item{mix2genos}{named list with one component per mixture, each component being a vector of genotype names; see \code{\link{getMixtureList}}}

\item{conf_level}{confidence level for the uncertainty intervals}

\item{plotDiag}{if TRUE, diagnostic plots}
}
\value{
data frame with the inference summaries of RY per mixture
}
\description{
Computes the relative yields (RYs) per replicate (e.g., blocks), then fits a linear model correcting for this rep effect, and quantifies the uncertainty around it.
}
\examples{
(dat <- data.frame(ID=c("geno1", "geno1",
                        "geno2", "geno2",
                        "mixg1g2", "mixg1g2", "mixg1g2", "mixg1g2"),
                   focal=c("geno1", "geno1",
                           "geno2", "geno2",
                           "geno1", "geno2",
                           "geno1", "geno2"),
                   prop=c(1, 1, 1, 1, 0.5, 0.5, 0.5, 0.5),
                   block=c("A","B","A","B","A","A","B","B"),
                   yield=c(51,45, 39,43, 25,22,26,25)))
estimRYRep(dat, "yield", "ID", "focal", "block", list("mixg1g2"=c("geno1","geno2")))
}
\seealso{
\code{\link{estimRYMRep}}
}
\author{
Timothee Flutre [aut], Arnaud Gauffreteau [ctb]
}
