% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_designs_var_mix.R
\name{getDesignBinaryVarMix}
\alias{getDesignBinaryVarMix}
\title{Optimize design for binary varietal mixtures}
\usage{
getDesignBinaryVarMix(
  levGenos,
  nbMixes,
  seed = NULL,
  showPlots = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{levGenos}{character vector with the names of the genotypes}

\item{nbMixes}{number of mixtures}

\item{seed}{seed for the pseudo-random number generator}

\item{showPlots}{logical indicating if plots should be showed}

\item{verbose}{verbosity level}
}
\value{
list
}
\description{
Optimizes a design for binary varietal mixtures
}
\examples{
nbGenos <- 25
levGenos <- sprintf(fmt=paste0("geno\%0", floor(log10(nbGenos))+1, "i"),
                    1:nbGenos)
nbMixes <- 75 # only binary and balanced
design <- getDesignBinaryVarMix(levGenos, nbMixes, seed=12345)
plotDesignVarMix(design$graph, levGenos)
ggplot(design$entropies) + aes(x=iteration, y=entropy) + geom_point() + geom_line() +
  geom_hline(yintercept = design$max_ent) + theme_bw()
}
\author{
Timothee Flutre
}
