% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ability.R
\name{getMixtureList}
\alias{getMixtureList}
\title{Reformat composition of mixtures into a list}
\usage{
getMixtureList(mixtures, sep = NULL)
}
\arguments{
\item{mixtures}{vector, matrix, data.frame or list containing of mixtures}

\item{sep}{required only if \code{mixtures} is a vector; will be given to \code{strsplit}}
}
\value{
list of vectors, with one component per mixture, the vector containing the mixture components
}
\description{
Reformats composition of mixtures into a list.
}
\examples{
nbGenos <- 25
levGenos <- sprintf(fmt=paste0("geno\%0", floor(log10(nbGenos))+1, "i"),
                    1:nbGenos)
nbMixes <- 75 # only binary and balanced
design <- getDesignBinaryVarMix(levGenos, nbMixes, seed=12345)
mixtures <- getMixtureList(design$combs)
str(mixtures, list.len=10)
}
\seealso{
\code{\link{getMixturesPerGeno}}
}
\author{
Timothee Flutre
}
