% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ability.R
\name{getMixturesPerGeno}
\alias{getMixturesPerGeno}
\title{Get mixture(s) per genotype}
\usage{
getMixturesPerGeno(mix2genos)
}
\arguments{
\item{mix2genos}{list with one component per mixture, listing the genotypes it contains (output of \code{\link{getMixtureList}})}
}
\value{
list with one vector per genotype
}
\description{
Returns the list of mixtures per genotype.
}
\examples{
nbGenos <- 25
levGenos <- sprintf(fmt=paste0("geno\%0", floor(log10(nbGenos))+1, "i"),
                    1:nbGenos)
nbMixes <- 75 # only binary and balanced
design <- getDesignBinaryVarMix(levGenos, nbMixes, seed=12345)
mixtures <- getMixtureList(design$combs)
str(mixtures, list.len=10)
geno2mixes <- getMixturesPerGeno(mixtures)
geno2mixes[c(1,2)]
table(sapply(geno2mixes, length))
}
\seealso{
\code{\link{getMixtureList}}
}
\author{
Timothee Flutre
}
