% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear_algebra.R
\name{imageMat}
\alias{imageMat}
\title{Image of a matrix}
\usage{
imageMat(mat, title, col, breaks)
}
\arguments{
\item{mat}{matrix}

\item{title}{optional title of the plot; if missing, it will be the matrix name and dimensions}

\item{col}{optional vector of colors; if missing, it will be a grey color gradient}

\item{breaks}{optional vector of breaks to go with col}
}
\value{
nothing
}
\description{
Plots an image of a matrix.
}
\examples{
M <- diag(c(rep(2, 3), rep(15, 5)))
imageMat(M) # see how the the "2" values are ignored!
imageMat(M, col=c("lightgrey","red","blue"), breaks=c(-1,1,3,20))
}
\author{
Timothee Flutre
}
