% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stat.R
\name{infoCriterion}
\alias{infoCriterion}
\title{Information criterion}
\usage{
infoCriterion(k, lnLmax, n = NULL, type = "AIC")
}
\arguments{
\item{k}{number of parameters, sometimes also called "effective degrees of freedom"}

\item{lnLmax}{value of the log-likelihood when maximized}

\item{n}{sample size}

\item{type}{specific information criterion to be returned}
}
\value{
numeric with attributes \code{k} and \code{n} (if not NULL)
}
\description{
Returns an information criterion, among AIC, AICc and BIC.
Caution: choosing \code{k} and \code{n} may not be straightforward for certain models, such as mixed models.
}
\author{
Timothee Flutre
}
