% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear_algebra.R
\name{invvec}
\alias{invvec}
\title{Inverse vec operator}
\usage{
invvec(x, n_col, n_row = NULL, sep = NULL)
}
\arguments{
\item{x}{vector}

\item{n_col}{number of columns of the output matrix}

\item{n_row}{number of rows of the output matrix}

\item{sep}{separator used to retrieve row and column names, only if \code{x} has names}
}
\value{
matrix
}
\description{
Applies the inverse vec operator to a vector.
}
\examples{
mat <- matrix(c(1, 2, 3,
                4, 5, 6),
              nrow = 2, ncol = 3, byrow = TRUE,
              dimnames = list(as.character(1:2), letters[1:3]))
mat
(x <- vec(mat))
invvec(x, n_col = 3, sep = "-")
}
\seealso{
\code{\link{vec}}
}
\author{
Timothee Flutre
}
