% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sowing.R
\name{mixSowingWeight}
\alias{mixSowingWeight}
\title{Sowing}
\usage{
mixSowingWeight(pureSowingWeights, stands)
}
\arguments{
\item{pureSowingWeights}{vector which names are genotypes and values are sowing weights in pure stands}

\item{stands}{vector which names are stand identifiers and values are genotype(s), a single one if it is a pure stand, several ones if it is a mixed stand (separated by a dash "-", e.g., "var1-var8")}
}
\value{
list which components are stands and values are named vectors with sowing weight per genotype for each stand
}
\description{
Computes the seed weight per genotype for each stand, assuming equal sowing proportions.
}
\examples{
sowingArea <- 9.52
sowingDensity <- 160
(tmp <- nbSeedsToSownInPure(sowingArea, sowingDensity))
TKWs <- c("var1"=38.605, "var2"=40.051, "var6"=36.251, "var8"=33.368)
pureSowingWeights <- TKWs * tmp
stands <- c("var1"="var1", "mix8"="var1-var2", "mix34"="var1-var8",
            "mix50"="var1-var6-var8")
mixSowingWeight(pureSowingWeights, stands)
}
\seealso{
\code{\link{nbSeedsToSownInPure}}
}
\author{
Timothee Flutre
}
