% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ability.R
\name{mkAllZSMA}
\alias{mkAllZSMA}
\title{Design matrices of SMAs for GMA-SMA models}
\usage{
mkAllZSMA(
  df,
  col,
  sep,
  models = c("2p", "2", "2pp", "3", "3p"),
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{col}{name of the column containing the names of pure and mixed stands}

\item{sep}{separator used to separate the genotype names in \code{"col"}}

\item{models}{vector of model identifiers; see the vignette}

\item{verbose}{verbosity level}
}
\value{
list of SMA design matrices
}
\description{
Makes the design matrices with the contribution of specific mixing abilities (SMAs).
More details in the help of \code{\link{mkZSMA}} and in the vignette.
}
\examples{
dat <- data.frame(mix=paste0("mix", 1:3),
                  varieties=c("var3","var1-var3","var1-var2"),
                  pheno=c(10, 7, 9))
(listZSMA <- mkAllZSMA(df=dat, col="varieties", sep="-"))
}
\author{
Timothee Flutre
}
