% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ability.R
\name{mkZGMA}
\alias{mkZGMA}
\title{Design matrix of GMAs for GMA-SMA models}
\usage{
mkZGMA(df, col, sep = ",")
}
\arguments{
\item{df}{data frame}

\item{col}{name of the column containing the names of pure and mixed stands}

\item{sep}{separator used to separate the genotype names in \code{"col"}}
}
\value{
matrix
}
\description{
Makes a design matrix with the contribution of general mixing abilities (GMAs).
The design is based on the names of pure and mixed stands, in which the genotype names are separated by a specific symbol.
}
\examples{
dat <- data.frame(mix=paste0("mix", 1:3),
                  varieties=c("var3","var1-var3","var1-var2"),
                  pheno=c(10, 7, 9))
(Z_GMA <- mkZGMA(df=dat, col="varieties", sep="-"))
}
\seealso{
\code{\link{mkZGMA}}, \code{\link{fitGMASMA}}
}
\author{
Timothee Flutre
}
