% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sowing.R
\name{nbSeedsToSownInPure}
\alias{nbSeedsToSownInPure}
\title{Sowing}
\usage{
nbSeedsToSownInPure(sowingArea = 8.4, sowingDensity = 160)
}
\arguments{
\item{sowingArea}{area to be sowned in square meters}

\item{sowingDensity}{number of seeds per square meter}
}
\value{
number of seeds (in thousands)
}
\description{
Computes the number of seeds of a given genotype to be sowned as pure stand based on a given sowing area and density.
}
\examples{
sowingArea <- 9.52
sowingDensity <- 160
nbSeedsToSownInPure(sowingArea, sowingDensity)
sowingArea <- 8.66
sowingDensity <- 200
nbSeedsToSownInPure(sowingArea, sowingDensity)
}
\seealso{
\code{\link{mixSowingWeight}}
}
\author{
Timothee Flutre
}
