% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stat.R
\name{normBiasError}
\alias{normBiasError}
\title{Normalized bias error}
\usage{
normBiasError(estim, truth, perc = TRUE)
}
\arguments{
\item{estim}{numeric vector of estimates}

\item{truth}{numeric vector of true values}

\item{perc}{logical; if TRUE, the return value will be in percentage}
}
\value{
numeric vector
}
\description{
Returns the normalized bias error (NBE).
}
\author{
Timothee Flutre
}
