% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TMB.R
\name{paramBoot4TMB}
\alias{paramBoot4TMB}
\title{Parametric bootstrap with TMB}
\usage{
paramBoot4TMB(fit, nb_boot = 5, mc.cores = 1)
}
\arguments{
\item{fit}{list corresponding to a model fitted with TMB}

\item{nb_boot}{number of parametric bootstraps to perform}

\item{mc.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (see the "parallel" package)}
}
\value{
list with as many components as \code{nb_boot}
}
\description{
Performs parametric bootstrap with TMB and \code{\link{nlminb}} on a fitted model to allow uncertainty quantification.
}
\examples{
## see the example of `fitDBVSBVinter`
## then run (slow if nb_boot is high!): paramBoot4TMB(fitTmb)
}
\author{
Jemay Salomon, Timothee Flutre
}
