% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_designs_crop_mix.R
\name{plotDesignCropMix}
\alias{plotDesignCropMix}
\title{Plot design for crop mixtures}
\usage{
plotDesignCropMix(graph, levGenos1, levGenos2, main = NULL)
}
\arguments{
\item{graph}{graph}

\item{levGenos1}{character vector with the names of the genotypes of the first species}

\item{levGenos2}{character vector with the names of the genotypes of the second species}

\item{main}{main title for the graph plot}
}
\value{
nothing
}
\description{
Plots design for crop mixtures, as a graph and as a diallel.
}
\examples{
nbGenos1 <- 30
levGenos1 <- sprintf(fmt=paste0("S1_\%0", floor(log10(nbGenos1))+1, "i"),
                     1:nbGenos1)
nbGenos2 <- 8
levGenos2 <- sprintf(fmt=paste0("S2_\%0", floor(log10(nbGenos2))+1, "i"),
                     1:nbGenos2)
nbMixes <- 60 # only binary and balanced
design <- getDesignBinaryCropMix(levGenos1, levGenos2, nbMixes, seed=12345)
plotDesignCropMix(design$graph, levGenos1, levGenos2)
}
\author{
Timothee Flutre
}
