% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_designs_var_mix.R
\name{plotDesignVarMix}
\alias{plotDesignVarMix}
\title{Plot design for varietal mixtures}
\usage{
plotDesignVarMix(
  graph,
  levGenos,
  main = NULL,
  subplots = c("graph", "diallel")
)
}
\arguments{
\item{graph}{graph}

\item{levGenos}{character vector with the names of the genotypes}

\item{main}{main title for the graph plot}

\item{subplots}{character vector indicating which object(s) to plot}
}
\value{
nothing
}
\description{
Plots design for varietal mixtures, as a graph and as a diallel.
}
\examples{
nbGenos <- 25
levGenos <- sprintf(fmt=paste0("geno\%0", floor(log10(nbGenos))+1, "i"),
                    1:nbGenos)
nbMixes <- 75 # only binary and balanced
design <- getDesignBinaryVarMix(levGenos, nbMixes, seed=12345)
plotDesignVarMix(design$graph, levGenos)
}
\author{
Timothee Flutre
}
