% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_designs_common.R
\name{plotDiallel}
\alias{plotDiallel}
\title{Plot diallel}
\usage{
plotDiallel(diallel, main = NULL)
}
\arguments{
\item{diallel}{matrix}

\item{main}{title}
}
\value{
nothing
}
\description{
Plots a diallel matrix.
}
\examples{
nbGenos <- 25
levGenos <- sprintf(fmt=paste0("geno\%0", floor(log10(nbGenos))+1, "i"),
                    1:nbGenos)
nbMixes <- 75 # only binary and balanced
design <- getDesignBinaryVarMix(levGenos, nbMixes, seed=12345)
plotDiallel(design$diallel)
}
\author{
Timothee Flutre
}
