% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{RYT}
\alias{RYT}
\title{Relative yield total (RYT)}
\usage{
RYT(
  dat,
  colIDstand = "ID",
  colIDfocal = "focal",
  colProp = "prop",
  colY = "yield"
)
}
\arguments{
\item{dat}{data.frame with one yield value per row and with at least four columns: the identifier of each stand, the identifier of the focal genotype or species in each stand, the sowing (plant) proportion of the focal in each stand, and the focal yield in each stand (see the example below); a stand with a single proportion equals to 1 will be interpreted as a monovarietal culture ("pure stands") and a mixture otherwise; note that contrary to \code{\link{RYP}}, the proportions will not be used for anything else}

\item{colIDstand}{column name for the stand identifiers}

\item{colIDfocal}{column name for the focal identifiers}

\item{colProp}{column name for the proportions}

\item{colY}{column name for the yield values}
}
\value{
input data.frame with an extra column named "RYT"
}
\description{
Computes the "relative yield total" (RYT) of a mixture from \href{http://edepot.wur.nl/197683}{de Wit and Den Bergh (1965)} as the sum of all relative yields (RY) of the mixture.
It corresponds to equation 26 (based on 25a and not 25b!) in Weigelt and Jolliffe (2003, \doi{10.1046/j.1365-2745.2003.00805.x}), and is equivalent to the index called "land equivalent ratio" (LER) corresponding to equation 28 of the same paper.
Missing values are forbidden.
}
\examples{
(dat <- data.frame(ID=c("geno1", "geno2", "mixg1g2", "mixg1g2"),
                   focal=c("geno1", "geno2", "geno1", "geno2"),
                   prop=c(1, 1, 0.5, 0.5),
                   yield=c(50, 40, 25, 22)))
RYT(dat)
}
\seealso{
\code{\link{RYP}}, \code{\link{RY}}, \code{\link{RYM}}, \code{\link{LER}}
}
