% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBV_SBV_crop_mix_simul.R
\name{simulDBVSBVinter}
\alias{simulDBVSBVinter}
\title{Simulate w.r.t. a DBV-SBV model for interspecific mixtures}
\usage{
simulDBVSBVinter(
  GRMs,
  parsGenetics = list(mu = list(S1 = c(SC = 65, IC = 32), S2 = c(SC = 30, IC = 27)),
    sdBlocks = 4, CV_g = c(S1 = 0.08, S2 = 0.08), prop_var_SBV = c(S1 = 0.2, S2 = 0.2),
    prop_var_SIGV = c(S1 = 0.5, S2 = 0), cor_DBV_SBV = c(S1 = -0.9, S2 = -0.9),
    prop_var_DBVxSBV = c(S1 = 0, S2 = 0), use_GRM_for_DBVxSBV = TRUE, cor_DxS = 0, H2_SC
    = c(S1 = 0.7, S2 = 0.7), T2 = c(S1 = 0.7, S2 = 0.7), cor_err_IC = -0.2),
  parsDesign = list(nbBlocks = 2, nbPlotsPerBl = 500, nbYs = 10, sowingDensities =
    list(S1 = c(SC = 300, IC = 150), S2 = c(SC = 40, IC = 40)), testersS2 = TRUE,
    incomplete_balanced = TRUE),
  sep = "+",
  seed = NULL
)
}
\arguments{
\item{GRMs}{named list with a genomic relationship matrix per species, named S1 and S2; should be a diagonal for S2 if it is a tester (see \code{parsDesign})}

\item{parsGenetics}{list of genetic parameters}

\item{parsDesign}{list of design parameters}

\item{sep}{character separating the names of a mixture components}

\item{seed}{seed for the generation of pseudo-random numbers}
}
\value{
list
}
\description{
Simulates an incomplete (sparse) yet balanced, tester-based design with respect to a DBV-SBV model for interspecific mixtures.
}
\examples{
## simulate a data set with both sole crops and intercrops:
GRMs <- list("S1" = diag(100),
             "S2" = diag(2))
dimnames(GRMs$S1) <- list(paste0("gS1_", 1:100), paste0("gS1_", 1:100))
dimnames(GRMs$S2) <- list(paste0("gS2_", 1:2), paste0("gS2_", 1:2))
out <- simulDBVSBVinter(GRMs)
names(out)
str(out$datW)
str(out$datL)

## see the third vignette for more details
}
\seealso{
\code{\link{fitDBVSBVinter}}
}
\author{
Timothee Flutre
}
