% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ability.R
\name{summarizeGMASMAs}
\alias{summarizeGMASMAs}
\title{Model summary}
\usage{
summarizeGMASMAs(fits)
}
\arguments{
\item{fits}{named list of "merMod" objects returned by \code{\link[lme4]{lmer}}, as is done by \code{\link{fitGMASMA}} with \code{pkg="lme4"}}
}
\value{
matrix
}
\description{
Summarizes the GMA-SMA models.
}
\examples{
## see the first vignette
}
\seealso{
\code{\link{fitGMASMA}}
}
\author{
Timothee Flutre
}
